ratatui::backend

Struct CrosstermBackend

pub struct CrosstermBackend<W>
where
    W: Write,
{ /* private fields */ }

A Backend implementation that uses Crossterm to render to the terminal.

The CrosstermBackend struct is a wrapper around a writer implementing Write, which is used to send commands to the terminal. It provides methods for drawing content, manipulating the cursor, and clearing the terminal screen.

Most applications should not call the methods on CrosstermBackend directly, but will instead use the Terminal struct, which provides a more ergonomic interface.

Usually applications will enable raw mode and switch to alternate screen mode after creating a CrosstermBackend. This is done by calling crossterm::terminal::enable_raw_mode and crossterm::terminal::EnterAlternateScreen (and the corresponding disable/leave functions when the application exits). This is not done automatically by the backend because it is possible that the application may want to use the terminal for other purposes (like showing help text) before entering alternate screen mode.

Example

use std::io::{stderr, stdout};

use crossterm::ExecutableCommand;
use crossterm::terminal::{
    EnterAlternateScreen, LeaveAlternateScreen, disable_raw_mode, enable_raw_mode,
};
use ratatui::Terminal;
use ratatui::backend::CrosstermBackend;

let mut backend = CrosstermBackend::new(stdout());
// or
let backend = CrosstermBackend::new(stderr());
let mut terminal = Terminal::new(backend)?;

enable_raw_mode()?;
stdout().execute(EnterAlternateScreen)?;

terminal.clear()?;
terminal.draw(|frame| {
    // -- snip --
})?;

stdout().execute(LeaveAlternateScreen)?;
disable_raw_mode()?;

See the the Examples directory for more examples. See the backend module documentation for more details on raw mode and alternate screen.

Implementations

impl<W> CrosstermBackend<W>
where W: Write,

pub const fn new(writer: W) -> CrosstermBackend<W>

Creates a new CrosstermBackend with the given writer.

Most applications will use either stdout or stderr as writer. See the FAQ to determine which one to use.

Example

use std::io::stdout;

use ratatui::backend::CrosstermBackend;

let backend = CrosstermBackend::new(stdout());