Struct ratatui::backend::TermionBackend

pub struct TermionBackend<W>where
    W: Write,{ /* private fields */ }

A backend that uses the Termion library to draw content, manipulate the cursor, and clear the terminal screen.

Example

use ratatui::backend::{Backend, TermionBackend};

let stdout = std::io::stdout();
let mut backend = TermionBackend::new(stdout);
backend.clear()?;

Implementations

impl<W> TermionBackend<W>where W: Write,

pub fn new(stdout: W) -> TermionBackend<W>

Creates a new Termion backend with the given output.