ratatui::prelude::text

Trait ToSpan

pub trait ToSpan {
    // Required method
    fn to_span(&self) -> Span<'_>;
}

A trait for converting a value to a Span.

This trait is automatically implemented for any type that implements the Display trait. As such, ToSpan shouln’t be implemented directly: Display should be implemented instead, and you get the ToSpan implementation for free.

Required Methods

fn to_span(&self) -> Span<'_>

Converts the value to a Span.