ratatui::prelude

Type Alias Alignment

pub type Alignment = HorizontalAlignment;

A type alias for HorizontalAlignment.

Prior to Ratatui 0.30.0, HorizontalAlignment was named Alignment. This alias is provided for backwards compatibility. Because this type is used almost everywhere in Ratatui related apps and libraries, it’s unlikely that this alias will be removed in the future.

Aliased Type

enum Alignment {
    Left,
    Center,
    Right,
}

Variants

Left

Center

Right