ratatui::style::palette::material

Struct AccentedPalette

pub struct AccentedPalette {

    pub c50: Color,
    pub c100: Color,
    pub c200: Color,
    pub c300: Color,
    pub c400: Color,
    pub c500: Color,
    pub c600: Color,
    pub c700: Color,
    pub c800: Color,
    pub c900: Color,
    pub a100: Color,
    pub a200: Color,
    pub a400: Color,
    pub a700: Color,

}

A palette of colors for use in Material design with accent colors

This is a collection of colors that are used in Material design. They consist of a set of colors from 50 to 900, and a set of accent colors from 100 to 700.

Fields
c50: Colorc100: Colorc200: Colorc300: Colorc400: Colorc500: Colorc600: Colorc700: Colorc800: Colorc900: Colora100: Colora200: Colora400: Colora700: Color
Implementations

impl AccentedPalette

pub const fn from_variants(variants: [u32; 14]) -> Self

Create a new AccentedPalette from the given variants

The variants should be in the format [0x00RRGGBB, โฆ]