ratatui::style::palette

Module tailwind

Represents the Tailwind CSS default color palette.

There are 22 palettes. Each palette has 11 colors, with variants from 50 to 950. Black and White are also included for completeness and to avoid being affected by any terminal theme that might be in use.

C50

C100

C200

C300

C400

C500

C600

C700

C800

C900

C950

SLATE

GRAY

ZINC

NEUTRAL

STONE

RED

ORANGE

AMBER

YELLOW

LIME

GREEN

EMERALD

TEAL

CYAN

SKY

BLUE

INDIGO

VIOLET

PURPLE

FUCHSIA

PINK

BLACK

WHITE

Example

use ratatui_core::style::Color;
use ratatui_core::style::palette::tailwind::{BLUE, RED};

assert_eq!(RED.c500, Color::Rgb(239, 68, 68));
assert_eq!(BLUE.c500, Color::Rgb(59, 130, 246));

Structs

Palette

Constants

AMBER

BLACK

BLUE

CYAN

EMERALD

FUCHSIA

GRAY

GREEN

INDIGO

LIME

NEUTRAL

ORANGE

PINK

PURPLE

RED

ROSE

SKY

SLATE

STONE

TEAL

VIOLET

WHITE

YELLOW

ZINC