ratatui::symbols::merge

Enum MergeStrategy

pub enum MergeStrategy {
    Replace,
    Exact,
    Fuzzy,
}

A strategy for merging two symbols into one.

This enum defines how two symbols should be merged together, allowing for different behaviors when combining symbols, such as replacing the previous symbol, merging them if an exact match exists, or using a fuzzy match to find the closest representation.

This is useful for collapsing borders in layouts, where multiple symbols may need to be combined to create a single, coherent border representation.

Not all combinations of box drawing symbols can be represented as a single unicode character, as many of them are not defined in the Box Drawing Unicode block. This means that some merging strategies will not yield a valid unicode character. The MergeStrategy::Replace strategy will be used as a fallback in such cases, replacing the previous symbol with the next one.

Specifically, the following combinations of box drawing symbols are not defined in the Box Drawing Unicode block:

Combining any dashed segments with any non dashed segments (e.g. ╎ with ─ or ━).

Combining any rounded segments with any other segments (e.g. ╯ with ─ or ━).

Combining any double segments with any thick segments (e.g. ═ with ┃ or ━).

Combining some double segments with some plain segments (e.g. ┐ with ╔).

The merging strategies include:

Self::Replace: Replaces the previous symbol with the next one.

Self::Exact: Merges symbols only if an exact composite unicode character exists, falling back to Self::Replace if not.

Self::Fuzzy: Merges symbols even if an exact composite unicode character doesn’t exist, using the closest match, and falling back to Self::Exact if necessary.

See Cell::merge_symbol for how to use this strategy in practice, and Block::merge_borders for a more concrete example of merging borders in a layout.

Examples

assert_eq!(MergeStrategy::Replace.merge("│", "━"), "━");
assert_eq!(MergeStrategy::Exact.merge("│", "─"), "┼");
assert_eq!(MergeStrategy::Fuzzy.merge("┘", "╔"), "╬");

Variants

Replace

Replaces the previous symbol with the next one.

This strategy simply replaces the previous symbol with the next one, without attempting to merge them. This is useful when you want to ensure that the last rendered symbol takes precedence over the previous one, regardless of their compatibility.

The following diagram illustrates how this would apply to several overlapping blocks where the thick bordered blocks are rendered last, replacing the previous symbols:

┌───┐    ┌───┐  ┌───┏━━━┓┌───┐
│   │    │   │  │   ┃   ┃│   │
│   │    │ ┏━━━┓│   ┃   ┃│   │
│   │    │ ┃ │ ┃│   ┃   ┃│   │
└───┏━━━┓└─┃─┘ ┃└───┗━━━┛┏━━━┓
    ┃   ┃  ┃   ┃         ┃   ┃
    ┃   ┃  ┗━━━┛         ┃   ┃
    ┃   ┃                ┃   ┃
    ┗━━━┛                ┗━━━┛

Example

let strategy = MergeStrategy::Replace;
assert_eq!(strategy.merge("│", "━"), "━");

Exact

Merges symbols only if an exact composite unicode character exists.

This strategy attempts to merge two symbols into a single composite unicode character if the exact representation exists. If the required unicode symbol does not exist, it falls back to MergeStrategy::Replace, replacing the previous symbol with the next one.

The following diagram illustrates how this would apply to several overlapping blocks where the thick bordered blocks are rendered last, merging the previous symbols into a single composite character. All combindations of the plain and thick segments exist, so these symbols can be merged into a single character:

┌───┐    ┌───┐  ┌───┲━━━┓┌───┐
│   │    │   │  │   ┃   ┃│   │
│   │    │ ┏━┿━┓│   ┃   ┃│   │
│   │    │ ┃ │ ┃│   ┃   ┃│   │
└───╆━━━┓└─╂─┘ ┃└───┺━━━┛┢━━━┪
    ┃   ┃  ┃   ┃         ┃   ┃
    ┃   ┃  ┗━━━┛         ┃   ┃
    ┃   ┃                ┃   ┃
    ┗━━━┛                ┗━━━┛

The following diagram illustrates how this would apply to several overlapping blocks where the characters don’t have a composite unicode character, so the previous symbols are replaced by the next one:

┌───┐    ┌───┐  ┌───╔═══╗┌───┐
│   │    │   │  │   ║   ║│   │
│   │    │ ╔═╪═╗│   ║   ║│   │
│   │    │ ║ │ ║│   ║   ║│   │
└───╔═══╗└─╫─┘ ║└───╚═══╝╔═══╗
    ║   ║  ║   ║         ║   ║
    ║   ║  ╚═══╝         ║   ║
    ║   ║                ║   ║
    ╚═══╝                ╚═══╝
┌───┐    ┌───┐  ┌───╭───╮┌───┐
│   │    │   │  │   │   ││   │
│   │    │ ╭─┼─╮│   │   ││   │
│   │    │ │ │ ││   │   ││   │
└───╭───╮└─┼─┘ │└───╰───╯╭───╮
    │   │  │   │         │   │
    │   │  ╰───╯         │   │
    │   │                │   │
    ╰───╯                ╰───╯

Example

let strategy = MergeStrategy::Exact;
assert_eq!(strategy.merge("│", "━"), "┿"); // exact match exists
assert_eq!(strategy.merge("┘", "╔"), "╔"); // no exact match, falls back to Replace

Fuzzy

Merges symbols even if an exact composite unicode character doesn’t exist, using the closest match.

If required unicode symbol exists, acts exactly like MergeStrategy::Exact, if not, the following rules are applied:

There are no characters that combine dashed with plain / thick segments, so we replace dashed segments with plain and thick dashed segments with thick. The following diagram shows how this would apply to merging a block with thick dashed borders over a block with plain dashed borders:

┌╌╌╌┐    ┌╌╌╌┐  ┌╌╌╌┲╍╍╍┓┌╌╌╌┐
╎   ╎    ╎   ╎  ╎   ╏   ╏╎   ╎
╎   ╎    ╎ ┏╍┿╍┓╎   ╏   ╏╎   ╎
╎   ╎    ╎ ╏ ╎ ╏╎   ╏   ╏╎   ╎
└╌╌╌╆╍╍╍┓└╌╂╌┘ ╏└╌╌╌┺╍╍╍┛┢╍╍╍┪
    ╏   ╏  ╏   ╏         ╏   ╏
    ╏   ╏  ┗╍╍╍┛         ╏   ╏
    ╏   ╏                ╏   ╏
    ┗╍╍╍┛                ┗╍╍╍┛

There are no characters that combine rounded segments with other segments, so we replace rounded segments with plain. The following diagram shows how this would apply to merging a block with rounded corners over a block with plain corners:

┌───┐    ┌───┐  ┌───┬───╮┌───┐
│   │    │   │  │   │   ││   │
│   │    │ ╭─┼─╮│   │   ││   │
│   │    │ │ │ ││   │   ││   │
└───┼───╮└─┼─┘ │└───┴───╯├───┤
    │   │  │   │         │   │
    │   │  ╰───╯         │   │
    │   │                │   │
    ╰───╯                ╰───╯

There are no symbols that combine thick and double borders, so we replace all double segments with thick or all thick with double. The second symbol parameter takes precedence in choosing whether to use double or thick. The following diagram shows how this would apply to merging a block with double borders over a block with thick borders and then the reverse (merging a block with thick borders over a block with double borders):

┏━━━┓    ┏━━━┓  ┏━━━╦═══╗┏━━━┓
┃   ┃    ┃   ┃  ┃   ║   ║┃   ┃
┃   ┃    ┃ ╔═╬═╗┃   ║   ║┃   ┃
┃   ┃    ┃ ║ ┃ ║┃   ║   ║┃   ┃
┗━━━╬═══╗┗━╬━┛ ║┗━━━╩═══╝╠═══╣
    ║   ║  ║   ║         ║   ║
    ║   ║  ╚═══╝         ║   ║
    ║   ║                ║   ║
    ╚═══╝                ╚═══╝

╔═══╗    ╔═══╗  ╔═══┳━━━┓╔═══╗
║   ║    ║   ║  ║   ┃   ┃║   ║
║   ║    ║ ┏━╋━┓║   ┃   ┃║   ║
║   ║    ║ ┃ ║ ┃║   ┃   ┃║   ║
╚═══╋━━━┓╚═╋═╝ ┃╚═══┻━━━┛┣━━━┫
    ┃   ┃  ┃   ┃         ┃   ┃
    ┃   ┃  ┗━━━┛         ┃   ┃
    ┃   ┃                ┃   ┃
    ┗━━━┛                ┗━━━┛

Some combinations of double and plain don’t exist, so if the symbol is still unrepresentable, change all plain segments with double or all double with plain. The second symbol parameter takes precedence in choosing whether to use double or plain. The following diagram shows how this would apply to merging a block with double borders over a block with plain borders and then the reverse (merging a block with plain borders over a block with double borders):

┌───┐    ┌───┐  ┌───╦═══╗┌───┐
│   │    │   │  │   ║   ║│   │
│   │    │ ╔═╪═╗│   ║   ║│   │
│   │    │ ║ │ ║│   ║   ║│   │
└───╬═══╗└─╫─┘ ║└───╩═══╝╠═══╣
    ║   ║  ║   ║         ║   ║
    ║   ║  ╚═══╝         ║   ║
    ║   ║                ║   ║
    ╚═══╝                ╚═══╝
╔═══╗    ╔═══╗  ╔═══┬───┐╔═══╗
║   ║    ║   ║  ║   │   │║   ║
║   ║    ║ ┌─╫─┐║   │   │║   ║
║   ║    ║ │ ║ │║   │   │║   ║
╚═══┼───┐╚═╪═╝ │╚═══┴───┘├───┤
    │   │  │   │         │   │
    │   │  └───┘         │   │
    │   │                │   │
    └───┘                └───┘

Examples

let strategy = MergeStrategy::Fuzzy;

// exact matches are merged normally
assert_eq!(strategy.merge("┌", "┐"), "┬");

// dashed segments are replaced with plain
assert_eq!(strategy.merge("╎", "╍"), "┿");

// rounded segments are replaced with plain
assert_eq!(strategy.merge("┘", "╭"), "┼");

// double and thick segments are merged based on the second symbol
assert_eq!(strategy.merge("┃", "═"), "╬");
assert_eq!(strategy.merge("═", "┃"), "╋");

// combindations of double with plain that don't exist are merged based on the second symbol
assert_eq!(strategy.merge("┐", "╔"), "╦");
assert_eq!(strategy.merge("╔", "┐"), "┬");

Implementations

impl MergeStrategy

pub fn merge<'a>(self, prev: &'a str, next: &'a str) -> &'a str

Merges two symbols using this merge strategy.

This method takes two string slices representing the previous and next symbols, and returns a string slice representing the merged symbol based on the merge strategy.

If either of the symbols are not in the Box Drawing Unicode block, the next symbol is returned as is. If both symbols are valid, they are merged according to the rules defined in the MergeStrategy.

Most code using this method will use the Cell::merge_symbol method, which uses this method internally to merge the symbols of a cell.

Example

let strategy = MergeStrategy::Fuzzy;
assert_eq!(strategy.merge("┌", "┐"), "┬"); // merges to a single character
assert_eq!(strategy.merge("┘", "╭"), "┼"); // replaces rounded with plain
assert_eq!(strategy.merge("╎", "╍"), "┿"); // replaces dashed with plain
assert_eq!(strategy.merge("┐", "╔"), "╦"); // merges double with plain
assert_eq!(strategy.merge("╔", "┐"), "┬"); // merges plain with double