Struct ratatui::text::Line

pub struct Line<'a> {
    pub spans: Vec<Span<'a>>,
    pub alignment: Option<Alignment>,
}

Fields
spans: Vec<Span<'a>>alignment: Option<Alignment>
Implementations

impl<'a> Line<'a>

pub fn styled<T>(content: T, style: Style) -> Line<'a>where T: Into<Cow<'a, str>>,

Create a line with a style.

Examples

let style = Style::default().fg(Color::Yellow).add_modifier(Modifier::ITALIC);
Line::styled("My text", style);
Line::styled(String::from("My text"), style);

pub fn width(&self) -> usize

Returns the width of the underlying string.

Examples

let line = Line::from(vec![
    Span::styled("My", Style::default().fg(Color::Yellow)),
    Span::raw(" text"),
]);
assert_eq!(7, line.width());

pub fn styled_graphemes( &'a self, base_style: Style ) -> impl Iterator<Item = StyledGrapheme<'a>>

Returns an iterator over the graphemes held by this line.

base_style is the Style that will be patched with each grapheme Style to get the resulting Style.

Examples

let line = Line::styled("Text", Style::default().fg(Color::Yellow));
let style = Style::default().fg(Color::Green).bg(Color::Black);
assert_eq!(
    line.styled_graphemes(style).collect::<Vec<StyledGrapheme>>(),
    vec![
        StyledGrapheme::new("T", Style::default().fg(Color::Yellow).bg(Color::Black)),
        StyledGrapheme::new("e", Style::default().fg(Color::Yellow).bg(Color::Black)),
        StyledGrapheme::new("x", Style::default().fg(Color::Yellow).bg(Color::Black)),
        StyledGrapheme::new("t", Style::default().fg(Color::Yellow).bg(Color::Black)),
    ]
);

pub fn patch_style(&mut self, style: Style)

Patches the style of each Span in an existing Line, adding modifiers from the given style.

Examples

let style = Style::default().fg(Color::Yellow).add_modifier(Modifier::ITALIC);
let mut raw_line = Line::from(vec![
    Span::raw("My"),
    Span::raw(" text"),
]);
let mut styled_line = Line::from(vec![
    Span::styled("My", style),
    Span::styled(" text", style),
]);

assert_ne!(raw_line, styled_line);

raw_line.patch_style(style);
assert_eq!(raw_line, styled_line);

pub fn reset_style(&mut self)

Resets the style of each Span in the Line. Equivalent to calling patch_style(Style::reset()).

Examples

let mut line = Line::from(vec![
    Span::styled("My", Style::default().fg(Color::Yellow)),
    Span::styled(" text", Style::default().add_modifier(Modifier::BOLD)),
]);

line.reset_style();
assert_eq!(Style::reset(), line.spans[0].style);
assert_eq!(Style::reset(), line.spans[1].style);

pub fn alignment(self, alignment: Alignment) -> Self

Sets the target alignment for this line of text. Defaults to: None, meaning the alignment is determined by the rendering widget.

Examples

let mut line = Line::from("Hi, what's up?");
assert_eq!(None, line.alignment);
assert_eq!(Some(Alignment::Right), line.alignment(Alignment::Right).alignment)