Struct ratatui::text::Spans

pub struct Spans<'a>(pub Vec<Span<'a>>);

๐Deprecated: Use ratatui::text::Line instead

A string composed of clusters of graphemes, each with their own style.

Spans has been deprecated in favor of Line, and will be removed in the future. All methods that accept Spans have been replaced with methods that accept Into<Line<โa>> (which is implemented on Spans) to allow users of this crate to gradually transition to Line.

Tuple Fields
0: Vec<Span<'a>>
๐Deprecated: Use ratatui::text::Line instead

Implementations

impl<'a> Spans<'a>

pub fn width(&self) -> usize

Returns the width of the underlying string.

Examples

let spans = Spans::from(vec![
    Span::styled("My", Style::default().fg(Color::Yellow)),
    Span::raw(" text"),
]);
assert_eq!(7, spans.width());

pub fn patch_style(&mut self, style: Style)

Patches the style of each Span in an existing Spans, adding modifiers from the given style.

Examples

let style = Style::default().fg(Color::Yellow).add_modifier(Modifier::ITALIC);
let mut raw_spans = Spans::from(vec![
    Span::raw("My"),
    Span::raw(" text"),
]);
let mut styled_spans = Spans::from(vec![
    Span::styled("My", style),
    Span::styled(" text", style),
]);

assert_ne!(raw_spans, styled_spans);

raw_spans.patch_style(style);
assert_eq!(raw_spans, styled_spans);

pub fn reset_style(&mut self)

Resets the style of each Span in the Spans. Equivalent to calling patch_style(Style::reset()).

Examples

let mut spans = Spans::from(vec![
    Span::styled("My", Style::default().fg(Color::Yellow)),
    Span::styled(" text", Style::default().add_modifier(Modifier::BOLD)),
]);

spans.reset_style();
assert_eq!(Style::reset(), spans.0[0].style);
assert_eq!(Style::reset(), spans.0[1].style);

pub fn alignment(self, alignment: Alignment) -> Line<'a>

Sets the target alignment for this line of text. Defaults to: None, meaning the alignment is determined by the rendering widget.

Examples

let mut line = Spans::from("Hi, what's up?").alignment(Alignment::Right);
assert_eq!(Some(Alignment::Right), line.alignment)