Struct ratatui::text::Text

pub struct Text<'a> {
    pub lines: Vec<Line<'a>>,
}

A string split over multiple lines where each line is composed of several clusters, each with their own style.

A Text, like a Span, can be constructed using one of the many From implementations or via the Text::raw and Text::styled methods. Helpfully, Text also implements core::iter::Extend which enables the concatenation of several Text blocks.

use ratatui::prelude::*;

let style = Style::default()
    .fg(Color::Yellow)
    .add_modifier(Modifier::ITALIC);

// An initial two lines of `Text` built from a `&str`
let mut text = Text::from("The first line\nThe second line");
assert_eq!(2, text.height());

// Adding two more unstyled lines
text.extend(Text::raw("These are two\nmore lines!"));
assert_eq!(4, text.height());

// Adding a final two styled lines
text.extend(Text::styled("Some more lines\nnow with more style!", style));
assert_eq!(6, text.height());

Fields
lines: Vec<Line<'a>>
Implementations

impl<'a> Text<'a>

pub fn raw<T>(content: T) -> Text<'a>where T: Into<Cow<'a, str>>,

Create some text (potentially multiple lines) with no style.

Examples

Text::raw("The first line\nThe second line");
Text::raw(String::from("The first line\nThe second line"));

pub fn styled<T>(content: T, style: Style) -> Text<'a>where T: Into<Cow<'a, str>>,

Create some text (potentially multiple lines) with a style.

Examples

let style = Style::default()
    .fg(Color::Yellow)
    .add_modifier(Modifier::ITALIC);
Text::styled("The first line\nThe second line", style);
Text::styled(String::from("The first line\nThe second line"), style);

pub fn width(&self) -> usize

Returns the max width of all the lines.

Examples

let text = Text::from("The first line\nThe second line");
assert_eq!(15, text.width());

pub fn height(&self) -> usize

Returns the height.

Examples

let text = Text::from("The first line\nThe second line");
assert_eq!(2, text.height());

pub fn patch_style(&mut self, style: Style)

Patches the style of each line in an existing Text, adding modifiers from the given style.

Examples

let style = Style::default()
    .fg(Color::Yellow)
    .add_modifier(Modifier::ITALIC);
let mut raw_text = Text::raw("The first line\nThe second line");
let styled_text = Text::styled(String::from("The first line\nThe second line"), style);
assert_ne!(raw_text, styled_text);

raw_text.patch_style(style);
assert_eq!(raw_text, styled_text);

pub fn reset_style(&mut self)

Resets the style of the Text. Equivalent to calling patch_style(Style::reset()).

Examples

let style = Style::default()
    .fg(Color::Yellow)
    .add_modifier(Modifier::ITALIC);
let mut text = Text::styled("The first line\nThe second line", style);

text.reset_style();
for line in &text.lines {
    for span in &line.spans {
        assert_eq!(Style::reset(), span.style);
    }
}