ratatui

Type Alias DefaultTerminal

pub type DefaultTerminal = Terminal<CrosstermBackend<Stdout>>;

A type alias for the default terminal type.

This is a Terminal using the CrosstermBackend which writes to Stdout. This is a reasonable default for most applications. To use a different backend or output stream, instead use Terminal and a backend of your choice directly.

Aliased Type

struct DefaultTerminal { /* private fields */ }