Module ratatui::widgets::block

Elements related to the Block base widget.

This holds everything needed to display and configure a Block.

In its simplest form, a Block is a border around another widget. It can have a title and padding.

Re-exports

pub use self::title::Position;

pub use self::title::Title;

Modules

title

This module holds the Title element and its related configuration types. A title is a piece of Block configuration.

Structs

Block

Base widget to be used to display a box border around all upper level ones.

Padding

Defines the padding of a Block.

Enums

BorderType

The type of border of a Block.