ratatui::widgets::block

Enum BorderType

pub enum BorderType {
    Plain,
    Rounded,
    Double,
    Thick,
    QuadrantInside,
    QuadrantOutside,
}

The type of border of a Block.

See the borders method of Block to configure its borders.

Variants

Plain

A plain, simple border.

This is the default

Example

â”Œâ”€â”€â”€â”€â”€â”€â”€â”
â”‚       â”‚
â””â”€â”€â”€â”€â”€â”€â”€â”˜

Rounded

A plain border with rounded corners.

Example

â•­â”€â”€â”€â”€â”€â”€â”€â•®
â”‚       â”‚
â•°â”€â”€â”€â”€â”€â”€â”€â•¯

Double

A doubled border.

Note this uses one character that draws two lines.

Example

â•”â•â•â•â•â•â•â•â•—
â•‘       â•‘
â•šâ•â•â•â•â•â•â•â•

Thick

A thick border.

Example

â”â”â”â”â”â”â”â”â”“
â”ƒ       â”ƒ
â”—â”â”â”â”â”â”â”â”›

QuadrantInside

A border with a single line on the inside of a half block.

Example

â–—â–„â–„â–„â–„â–„â–„â–„â––
â–       â–Œ
â–       â–Œ
â–â–€â–€â–€â–€â–€â–€â–€â–˜

QuadrantOutside

A border with a single line on the outside of a half block.

Example

â–›â–€â–€â–€â–€â–€â–€â–€â–œ
â–Œ       â–
â–Œ       â–
â–™â–„â–„â–„â–„â–„â–„â–„â–Ÿ

Implementations

impl BorderType

pub const fn border_symbols(border_type: Self) -> Set

Convert this BorderType into the corresponding Set of border symbols.

pub const fn to_border_set(self) -> Set

Convert this BorderType into the corresponding Set of border symbols.