ratatui::widgets
Module calendar

A simple calendar widget. (feature: widget-calendar)

The Monthly widget will display a calendar for the month provided in display_date. Days are styled using the default style unless:

show_surrounding is set, then days not in the display_date month will use that style.

a style is returned by the DateStyler for the day

Monthly has several controls for what should be displayed

Structs

CalendarEventStore

A simple DateStyler based on a HashMap

Monthly

Display a month calendar for the month containing display_date

Traits

DateStyler

Provides a method for styling a given date. Monthly is generic on this trait, so any type that implements this trait can be used.