ratatui::widgets::calendar

Trait DateStyler

pub trait DateStyler {
    // Required method
    fn get_style(&self, date: Date) -> Style;
}

Provides a method for styling a given date. Monthly is generic on this trait, so any type that implements this trait can be used.

Required Methods

fn get_style(&self, date: Date) -> Style

Given a date, return a style for that date