ratatui::widgets::canvas

Struct Painter

pub struct Painter<'a, 'b> { /* private fields */ }

Painter is an abstraction over the Context that allows to draw shapes on the grid.

It is used by the Shape trait to draw shapes on the grid. It can be useful to think of this as similar to the Buffer struct that is used to draw widgets on the terminal.

Implementations

impl<'a, 'b> Painter<'a, 'b>

pub fn get_point(&self, x: f64, y: f64) -> Option<(usize, usize)>

Convert the (x, y) coordinates to location of a point on the grid

(x, y) coordinates are expressed in the coordinate system of the canvas. The origin is in the lower left corner of the canvas (unlike most other coordinates in Ratatui where the origin is the upper left corner). The x and y bounds of the canvas define the specific area of some coordinate system that will be drawn on the canvas. The resolution of the grid is used to convert the (x, y) coordinates to the location of a point on the grid.

The grid coordinates are expressed in the coordinate system of the grid. The origin is in the top left corner of the grid. The x and y bounds of the grid are always [0, width - 1] and [0, height - 1] respectively. The resolution of the grid is used to convert the (x, y) coordinates to the location of a point on the grid.

Examples

use ratatui::{
    symbols,
    widgets::canvas::{Context, Painter},
};

let mut ctx = Context::new(2, 2, [1.0, 2.0], [0.0, 2.0], symbols::Marker::Braille);
let mut painter = Painter::from(&mut ctx);

let point = painter.get_point(1.0, 0.0);
assert_eq!(point, Some((0, 7)));

let point = painter.get_point(1.5, 1.0);
assert_eq!(point, Some((1, 3)));

let point = painter.get_point(0.0, 0.0);
assert_eq!(point, None);

let point = painter.get_point(2.0, 2.0);
assert_eq!(point, Some((3, 0)));

let point = painter.get_point(1.0, 2.0);
assert_eq!(point, Some((0, 0)));

pub fn paint(&mut self, x: usize, y: usize, color: Color)

Paint a point of the grid

Example

use ratatui::{
    style::Color,
    symbols,
    widgets::canvas::{Context, Painter},
};

let mut ctx = Context::new(1, 1, [0.0, 2.0], [0.0, 2.0], symbols::Marker::Braille);
let mut painter = Painter::from(&mut ctx);
painter.paint(1, 3, Color::Red);