Struct ratatui::widgets::scrollbar::Scrollbar

pub struct Scrollbar<'a> { /* private fields */ }

Scrollbar widget for tui-rs library.

This widget can be used to display a scrollbar in a terminal user interface. The following components of the scrollbar are customizable in symbol and style.

<--▮------->
^  ^   ^   ^
│  │   │   └ end
│  │   └──── track
│  └──────── thumb
└─────────── begin

Implementations

impl<'a> Scrollbar<'a>

pub fn new(orientation: ScrollbarOrientation) -> Self

pub fn orientation(self, orientation: ScrollbarOrientation) -> Self

Sets the orientation of the scrollbar. Resets the symbols to DOUBLE_VERTICAL or DOUBLE_HORIZONTAL based on orientation

pub fn orientation_and_symbol( self, orientation: ScrollbarOrientation, set: Set ) -> Self

Sets the orientation and symbols for the scrollbar from a Set.

pub fn thumb_symbol(self, thumb_symbol: &'a str) -> Self

Sets the symbol that represents the thumb of the scrollbar.

pub fn thumb_style(self, thumb_style: Style) -> Self

Sets the style that represents the thumb of the scrollbar.

pub fn track_symbol(self, track_symbol: &'a str) -> Self

Sets the symbol that represents the track of the scrollbar.

pub fn track_style(self, track_style: Style) -> Self

Sets the style that is used for the track of the scrollbar.

pub fn begin_symbol(self, begin_symbol: Option<&'a str>) -> Self

Sets the symbol that represents the beginning of the scrollbar.

pub fn begin_style(self, begin_style: Style) -> Self

Sets the style that is used for the beginning of the scrollbar.

pub fn end_symbol(self, end_symbol: Option<&'a str>) -> Self

Sets the symbol that represents the end of the scrollbar.

pub fn end_style(self, end_style: Style) -> Self

Sets the style that is used for the end of the scrollbar.

pub fn symbols(self, symbol: Set) -> Self

Sets the symbols used for the various parts of the scrollbar from a Set.

<--▮------->
^  ^   ^   ^
│  │   │   └ end
│  │   └──── track
│  └──────── thumb
└─────────── begin

Only sets begin_symbol and end_symbol if they already contain a value. If begin_symbol and/or end_symbol were set to None explicitly, this function will respect that choice.

pub fn style(self, style: Style) -> Self

Sets the style used for the various parts of the scrollbar from a Style.

<--▮------->
^  ^   ^   ^
│  │   │   └ end
│  │   └──── track
│  └──────── thumb
└─────────── begin