Struct ratatui::widgets::Axis

pub struct Axis<'a> { /* private fields */ }

An X or Y axis for the chart widget

Implementations

impl<'a> Axis<'a>

pub fn title<T>(self, title: T) -> Axis<'a>where T: Into<TextLine<'a>>,

pub fn title_style(self, style: Style) -> Axis<'a>

๐Deprecated since 0.10.0: You should use styling capabilities of text::Line given as argument of the title method to apply styling to the title.

pub fn bounds(self, bounds: [f64; 2]) -> Axis<'a>

pub fn labels(self, labels: Vec<Span<'a>>) -> Axis<'a>

pub fn style(self, style: Style) -> Axis<'a>

pub fn labels_alignment(self, alignment: Alignment) -> Axis<'a>

Defines the alignment of the labels of the axis. The alignment behaves differently based on the axis:

Y-Axis: The labels are aligned within the area on the left of the axis

X-Axis: The first X-axis label is aligned relative to the Y-axis