Struct ratatui::widgets::ListItem

pub struct ListItem<'a> { /* private fields */ }

A single item in a List

The item’s height is defined by the number of lines it contains. This can be queried using ListItem::height. Similarly, ListItem::width will return the maximum width of all lines.

You can set the style of an item with ListItem::style or using the Stylize trait. This Style will be combined with the Style of the inner Text. The Style of the Text will be added to the Style of the ListItem.

You can also align a ListItem by aligning its underlying Text and Lines. For that, see Text::alignment and Line::alignment. On a multiline Text, one Line can override the alignment by setting it explicitly.

Examples

You can create ListItems from simple &str

let item = ListItem::new("Item 1");

Anything that can be converted to Text can be a ListItem.

let item1: ListItem = "Item 1".into();
let item2: ListItem = Line::raw("Item 2").into();

A ListItem styled with Stylize

let item = ListItem::new("Item 1").red().on_white();

If you need more control over the item’s style, you can explicitly style the underlying Text

let mut text = Text::default();
text.extend(["Item".blue(), Span::raw(" "), "1".bold().red()]);
let item = ListItem::new(text);

A right-aligned ListItem

ListItem::new(Text::from("foo").alignment(Alignment::Right));

Implementations

impl<'a> ListItem<'a>

pub fn new<T>(content: T) -> ListItem<'a>
where T: Into<Text<'a>>,

Creates a new ListItem

The content parameter accepts any value that can be converted into Text.

Examples

You can create ListItems from simple &str

let item = ListItem::new("Item 1");

Anything that can be converted to Text can be a ListItem.

let item1: ListItem = "Item 1".into();
let item2: ListItem = Line::raw("Item 2").into();

You can also create multilines item

let item = ListItem::new("Multi-line\nitem");

See also

List::new to create a list of items that can be converted to ListItem

pub fn style<S: Into<Style>>(self, style: S) -> ListItem<'a>

Sets the item style

style accepts any type that is convertible to Style (e.g. Style, Color, or your own type that implements Into<Style>).

This Style can be overridden by the Style of the Text content.

This is a fluent setter method which must be chained or used as it consumes self

Example

let item = ListItem::new("Item 1").style(Style::new().red().italic());

ListItem also implements the Styled trait, which means you can use style shorthands from the Stylize trait to set the style of the widget more concisely.

let item = ListItem::new("Item 1").red().italic();

pub fn height(&self) -> usize

Returns the item height

Examples

One line item

let item = ListItem::new("Item 1");
assert_eq!(item.height(), 1);

Two lines item (note the \n)

let item = ListItem::new("Multi-line\nitem");
assert_eq!(item.height(), 2);

pub fn width(&self) -> usize

Returns the max width of all the lines

Examples

let item = ListItem::new("12345");
assert_eq!(item.width(), 5);

let item = ListItem::new("12345\n1234567");
assert_eq!(item.width(), 7);