Struct ratatui::widgets::Paragraph

pub struct Paragraph<'a> { /* private fields */ }

A widget to display some text.

Example

use ratatui::{prelude::*, widgets::*};

let text = vec![
    Line::from(vec![
        Span::raw("First"),
        Span::styled("line",Style::new().green().italic()),
        ".".into(),
    ]),
    Line::from("Second line".red()),
    "Third line".into(),
];
Paragraph::new(text)
    .block(Block::new()
        .title("Paragraph")
        .borders(Borders::ALL))
    .style(Style::new().white().on_black())
    .alignment(Alignment::Center)
    .wrap(Wrap { trim: true });

Implementations

impl<'a> Paragraph<'a>

pub fn new<T>(text: T) -> Paragraph<'a>where T: Into<Text<'a>>,

Creates a new Paragraph widget with the given text.

The text parameter can be a Text or any type that can be converted into a Text. By default, the text is styled with Style::default(), not wrapped, and aligned to the left.

Examples

let paragraph = Paragraph::new("Hello, world!");
let paragraph = Paragraph::new(String::from("Hello, world!"));
let paragraph = Paragraph::new(Text::raw("Hello, world!"));
let paragraph = Paragraph::new(
    Text::styled("Hello, world!", Style::default()));
let paragraph = Paragraph::new(
    Line::from(vec!["Hello, ".into(), "world!".red()]));

pub fn block(self, block: Block<'a>) -> Paragraph<'a>

Surrounds the Paragraph widget with a Block.

Example

let paragraph = Paragraph::new("Hello, world!")
   .block(Block::default()
        .title("Paragraph")
        .borders(Borders::ALL));

pub fn style(self, style: Style) -> Paragraph<'a>

Sets the style of the entire widget.

This applies to the entire widget, including the block if one is present. Any style set on the block or text will be added to this style.

Example

let paragraph = Paragraph::new("Hello, world!")
   .style(Style::new().red().on_white());

pub fn wrap(self, wrap: Wrap) -> Paragraph<'a>

Sets the wrapping configuration for the widget.

See Wrap for more information on the different options.

Example

let paragraph = Paragraph::new("Hello, world!")
  .wrap(Wrap { trim: true });

pub fn scroll(self, offset: (u16, u16)) -> Paragraph<'a>

Set the scroll offset for the given paragraph

The scroll offset is a tuple of (y, x) offset. The y offset is the number of lines to scroll, and the x offset is the number of characters to scroll. The scroll offset is applied after the text is wrapped and aligned.

Note: the order of the tuple is (y, x) instead of (x, y), which is different from general convention across the crate.

For more information about future scrolling design and concerns, see RFC: Design of Scrollable Widgets on GitHub.

pub fn alignment(self, alignment: Alignment) -> Paragraph<'a>

Set the text alignment for the given paragraph

The alignment is a variant of the Alignment enum which can be one of Left, Right, or Center.

Example

let paragraph = Paragraph::new("Hello World")
    .alignment(Alignment::Center);