ratatui

Module backend

Re-exports for the backend implementations.

Structs

CrosstermBackend

A Backend implementation that uses Crossterm to render to the terminal.

TestBackend

A Backend implementation used for integration testing that renders to an memory buffer.

WindowSize

The window size in characters (columns / rows) as well as pixels.

Enums

ClearType

Enum representing the different types of clearing operations that can be performed on the terminal screen.

Traits

Backend

The Backend trait provides an abstraction over different terminal libraries. It defines the methods required to draw content, manipulate the cursor, and clear the terminal screen.

FromCrossterm

A trait for converting a Crossterm type to a Ratatui type.

IntoCrossterm

A trait for converting a Ratatui type to a Crossterm type.