Struct ratatui::backend::CrosstermBackend

pub struct CrosstermBackend<W: Write> { /* private fields */ }

A backend implementation using the crossterm crate.

The CrosstermBackend struct is a wrapper around a type implementing Write, which is used to send commands to the terminal. It provides methods for drawing content, manipulating the cursor, and clearing the terminal screen.

Example

use ratatui::backend::{Backend, CrosstermBackend};

let buffer = std::io::stdout();
let mut backend = CrosstermBackend::new(buffer);
backend.clear()?;

Implementations

impl<W> CrosstermBackend<W>where W: Write,

pub fn new(buffer: W) -> CrosstermBackend<W>

Creates a new CrosstermBackend with the given buffer.