ratatui::backend

Struct TestBackend

pub struct TestBackend { /* private fields */ }

A Backend implementation used for integration testing that renders to an memory buffer.

Note: that although many of the integration and unit tests in ratatui are written using this backend, it is preferable to write unit tests for widgets directly against the buffer rather than using this backend. This backend is intended for integration tests that test the entire terminal UI.

Example

use ratatui::backend::{Backend, TestBackend};

let mut backend = TestBackend::new(10, 2);
backend.clear()?;
backend.assert_buffer_lines(["          "; 2]);

Implementations

impl TestBackend

pub fn new(width: u16, height: u16) -> TestBackend

Creates a new TestBackend with the specified width and height.

pub fn with_lines<'line, Lines>(lines: Lines) -> TestBackend
where Lines: IntoIterator, <Lines as IntoIterator>::Item: Into<Line<'line>>,

Creates a new TestBackend with the specified lines as the initial screen state.

The backend’s screen size is determined from the initial lines.

pub const fn buffer(&self) -> &Buffer

Returns a reference to the internal buffer of the TestBackend.

pub const fn scrollback(&self) -> &Buffer

Returns a reference to the internal scrollback buffer of the TestBackend.

The scrollback buffer represents the part of the screen that is currently hidden from view, but that could be accessed by scrolling back in the terminal’s history. This would normally be done using the terminal’s scrollbar or an equivalent keyboard shortcut.

The scrollback buffer starts out empty. Lines are appended when they scroll off the top of the main buffer. This happens when lines are appended to the bottom of the main buffer using Backend::append_lines.

The scrollback buffer has a maximum height of u16::MAX. If lines are appended to the bottom of the scrollback buffer when it is at its maximum height, a corresponding number of lines will be removed from the top.

pub fn resize(&mut self, width: u16, height: u16)

Resizes the TestBackend to the specified width and height.

pub fn assert_buffer(&self, expected: &Buffer)

Asserts that the TestBackend’s buffer is equal to the expected buffer.

This is a shortcut for assert_eq!(self.buffer(), &expected).

Panics

When they are not equal, a panic occurs with a detailed error message showing the differences between the expected and actual buffers.

pub fn assert_scrollback(&self, expected: &Buffer)

Asserts that the TestBackend’s scrollback buffer is equal to the expected buffer.

This is a shortcut for assert_eq!(self.scrollback(), &expected).

Panics

When they are not equal, a panic occurs with a detailed error message showing the differences between the expected and actual buffers.

pub fn assert_scrollback_empty(&self)

Asserts that the TestBackend’s scrollback buffer is empty.

Panics

When the scrollback buffer is not equal, a panic occurs with a detailed error message showing the differences between the expected and actual buffers.

pub fn assert_buffer_lines<'line, Lines>(&self, expected: Lines)
where Lines: IntoIterator, <Lines as IntoIterator>::Item: Into<Line<'line>>,

Asserts that the TestBackend’s buffer is equal to the expected lines.

This is a shortcut for assert_eq!(self.buffer(), &Buffer::with_lines(expected)).

Panics

When they are not equal, a panic occurs with a detailed error message showing the differences between the expected and actual buffers.

pub fn assert_scrollback_lines<'line, Lines>(&self, expected: Lines)
where Lines: IntoIterator, <Lines as IntoIterator>::Item: Into<Line<'line>>,

Asserts that the TestBackend’s scrollback buffer is equal to the expected lines.

This is a shortcut for assert_eq!(self.scrollback(), &Buffer::with_lines(expected)).

Panics

When they are not equal, a panic occurs with a detailed error message showing the differences between the expected and actual buffers.

pub fn assert_cursor_position<P>(&mut self, position: P)
where P: Into<Position>,

Asserts that the TestBackend’s cursor position is equal to the expected one.

This is a shortcut for assert_eq!(self.get_cursor_position().unwrap(), expected).

Panics

When they are not equal, a panic occurs with a detailed error message showing the differences between the expected and actual position.