Struct ratatui::backend::TestBackend

pub struct TestBackend { /* private fields */ }

A backend used for the integration tests.

Example

use ratatui::{backend::{Backend, TestBackend}, buffer::Buffer};

let mut backend = TestBackend::new(10, 2);
backend.clear()?;
backend.assert_buffer(&Buffer::with_lines(vec!["          "; 2]));

Implementations

impl TestBackend

pub fn new(width: u16, height: u16) -> TestBackend

Creates a new TestBackend with the specified width and height.

pub fn buffer(&self) -> &Buffer

Returns a reference to the internal buffer of the TestBackend.

pub fn resize(&mut self, width: u16, height: u16)

Resizes the TestBackend to the specified width and height.

pub fn assert_buffer(&self, expected: &Buffer)

Asserts that the TestBackend’s buffer is equal to the expected buffer. If the buffers are not equal, a panic occurs with a detailed error message showing the differences between the expected and actual buffers.