Trait ratatui::backend::Backend

pub trait Backend {
    // Required methods
    fn draw<'a, I>(&mut self, content: I) -> Result<(), Error>
       where I: Iterator<Item = (u16, u16, &'a Cell)>;
    fn hide_cursor(&mut self) -> Result<(), Error>;
    fn show_cursor(&mut self) -> Result<(), Error>;
    fn get_cursor(&mut self) -> Result<(u16, u16), Error>;
    fn set_cursor(&mut self, x: u16, y: u16) -> Result<(), Error>;
    fn clear(&mut self) -> Result<(), Error>;
    fn size(&self) -> Result<Rect, Error>;
    fn flush(&mut self) -> Result<(), Error>;

    // Provided methods
    fn append_lines(&mut self, _n: u16) -> Result<()> { ... }
    fn clear_region(&mut self, clear_type: ClearType) -> Result<(), Error> { ... }
}

The Backend trait provides an abstraction over different terminal libraries. It defines the methods required to draw content, manipulate the cursor, and clear the terminal screen.

Required Methods

fn draw<'a, I>(&mut self, content: I) -> Result<(), Error>where I: Iterator<Item = (u16, u16, &'a Cell)>,

Draw the given content to the terminal screen.

The content is provided as an iterator over (u16, u16, &Cell) tuples, where the first two elements represent the x and y coordinates, and the third element is a reference to the Cell to be drawn.

fn hide_cursor(&mut self) -> Result<(), Error>

Hide the cursor on the terminal screen.

fn show_cursor(&mut self) -> Result<(), Error>

Show the cursor on the terminal screen.

fn get_cursor(&mut self) -> Result<(u16, u16), Error>

Get the current cursor position on the terminal screen.

fn set_cursor(&mut self, x: u16, y: u16) -> Result<(), Error>

Set the cursor position on the terminal screen to the given x and y coordinates.

fn clear(&mut self) -> Result<(), Error>

Clears the whole terminal screen

fn size(&self) -> Result<Rect, Error>

Get the size of the terminal screen as a Rect.

fn flush(&mut self) -> Result<(), Error>

Flush any buffered content to the terminal screen.

Provided Methods

fn append_lines(&mut self, _n: u16) -> Result<()>

Insert n line breaks to the terminal screen.

This method is optional and may not be implemented by all backends.

fn clear_region(&mut self, clear_type: ClearType) -> Result<(), Error>

Clears a specific region of the terminal specified by the ClearType parameter

This method is optional and may not be implemented by all backends.