ratatui

Module layout

Structs

Columns

An iterator over columns within a Rect.

Layout

A layout is a set of constraints that can be applied to a given area to split it into smaller ones.

Margin

Offset

Amounts by which to move a Rect.

Position

Position in the terminal

Positions

An iterator over positions within a Rect.

Rect

A Rectangular area.

Rows

An iterator over rows within a Rect.

Size

A simple size struct

Enums

Alignment

Constraint

A constraint that defines the size of a layout element.

Direction

Flex

Defines the options for layout flex justify content in a container.

Spacing

Represents the spacing between segments in a layout.