ratatui::layout

Struct Rect

pub struct Rect {
    pub x: u16,
    pub y: u16,
    pub width: u16,
    pub height: u16,
}

A Rectangular area.

A simple rectangle used in the computation of the layout and to give widgets a hint about the area they are supposed to render to.

Fields
x: u16

The x coordinate of the top left corner of the Rect.

y: u16

The y coordinate of the top left corner of the Rect.

width: u16

The width of the Rect.

height: u16

The height of the Rect.

Implementations

impl Rect

pub const ZERO: Self

A zero sized Rect at position 0,0

pub const fn new(x: u16, y: u16, width: u16, height: u16) -> Self

Creates a new Rect, with width and height limited to keep both bounds within u16.

If the width or height would cause the right or bottom coordinate to be larger than the maximum value of u16, the width or height will be clamped to keep the right or bottom coordinate within u16.

Examples

use ratatui::layout::Rect;

let rect = Rect::new(1, 2, 3, 4);

pub const fn area(self) -> u32

The area of the Rect. If the area is larger than the maximum value of u16, it will be clamped to u16::MAX.

pub const fn is_empty(self) -> bool

Returns true if the Rect has no area.

pub const fn left(self) -> u16

Returns the left coordinate of the Rect.

pub const fn right(self) -> u16

Returns the right coordinate of the Rect. This is the first coordinate outside of the Rect.

If the right coordinate is larger than the maximum value of u16, it will be clamped to u16::MAX.

pub const fn top(self) -> u16

Returns the top coordinate of the Rect.

pub const fn bottom(self) -> u16

Returns the bottom coordinate of the Rect. This is the first coordinate outside of the Rect.

If the bottom coordinate is larger than the maximum value of u16, it will be clamped to u16::MAX.

pub const fn inner(self, margin: Margin) -> Self

Returns a new Rect inside the current one, with the given margin on each side.

If the margin is larger than the Rect, the returned Rect will have no area.

pub fn offset(self, offset: Offset) -> Self

Moves the Rect without modifying its size.

Moves the Rect according to the given offset without modifying its width or height.

Positive x moves the whole Rect to the right, negative to the left.

Positive y moves the whole Rect to the bottom, negative to the top.

See Offset for details.

pub fn union(self, other: Self) -> Self

Returns a new Rect that contains both the current one and the given one.

pub fn intersection(self, other: Self) -> Self

Returns a new Rect that is the intersection of the current one and the given one.

If the two Rects do not intersect, the returned Rect will have no area.

pub const fn intersects(self, other: Self) -> bool

Returns true if the two Rects intersect.

pub const fn contains(self, position: Position) -> bool

Returns true if the given position is inside the Rect.

The position is considered inside the Rect if it is on the Rect’s border.

Examples

use ratatui::layout::{Position, Rect};

let rect = Rect::new(1, 2, 3, 4);
assert!(rect.contains(Position { x: 1, y: 2 }));

pub fn clamp(self, other: Self) -> Self

Clamp this Rect to fit inside the other Rect.

If the width or height of this Rect is larger than the other Rect, it will be clamped to the other Rect’s width or height.

If the left or top coordinate of this Rect is smaller than the other Rect, it will be clamped to the other Rect’s left or top coordinate.

If the right or bottom coordinate of this Rect is larger than the other Rect, it will be clamped to the other Rect’s right or bottom coordinate.

This is different from Rect::intersection because it will move this Rect to fit inside the other Rect, while Rect::intersection instead would keep this Rect’s position and truncate its size to only that which is inside the other Rect.

Examples

use ratatui::{layout::Rect, Frame};

let area = frame.area();
let rect = Rect::new(0, 0, 100, 100).clamp(area);

pub const fn rows(self) -> Rows

An iterator over rows within the Rect.

Example

use ratatui::{buffer::Buffer, layout::Rect, text::Line, widgets::Widget};

fn render(area: Rect, buf: &mut Buffer) {
    for row in area.rows() {
        Line::raw("Hello, world!").render(row, buf);
    }
}

pub const fn columns(self) -> Columns

An iterator over columns within the Rect.

Example

use ratatui::{
    buffer::Buffer,
    layout::Rect,
    widgets::{Block, Borders, Widget},
};
fn render(area: Rect, buf: &mut Buffer) {
    if let Some(left) = area.columns().next() {
        Block::new().borders(Borders::LEFT).render(left, buf);
    }
}

pub const fn positions(self) -> Positions

An iterator over the positions within the Rect.

The positions are returned in a row-major order (left-to-right, top-to-bottom).

Example

use ratatui::{buffer::Buffer, layout::Rect};

fn render(area: Rect, buf: &mut Buffer) {
    for position in area.positions() {
        buf[(position.x, position.y)].set_symbol("x");
    }
}

pub const fn as_position(self) -> Position

Returns a Position with the same coordinates as this Rect.

Examples

use ratatui::layout::Rect;

let rect = Rect::new(1, 2, 3, 4);
let position = rect.as_position();

pub const fn as_size(self) -> Size

Converts the Rect into a size struct.