ratatui

Module prelude

A prelude for conveniently writing applications using this library.

The prelude module is no longer used universally in Ratatui, as it can make it harder to distinguish between library and non-library types, especially when viewing source code outside of an IDE (such as on GitHub or in a git diff). For more details and user feedback, see Issue #1150. However, the prelude is still available for backward compatibility and for those who prefer to use it.

Examples

use ratatui::prelude::*;

Aside from the main types that are used in the library, this prelude also re-exports several modules to make it easy to qualify types that would otherwise collide. E.g.:

use ratatui::prelude::*;
use ratatui::widgets::*;

#[derive(Debug, Default, PartialEq, Eq)]
struct Line;

assert_eq!(Line::default(), Line);
assert_eq!(text::Line::default(), ratatui::text::Line::from(vec![]));

Modules

backend

This module provides the backend implementations for different terminal libraries.

buffer

A module for the Buffer and Cell types.

layout

Layout and positioning in terminal user interfaces.

style

style contains the primitives used to control how your user interface will look.

symbols

Symbols and markers for drawing various widgets.

text

Primitives for styled text.

Structs

Buffer

A buffer that maps to the desired content of the terminal after the draw call

CrosstermBackend

A Backend implementation that uses Crossterm to render to the terminal.

Frame

A consistent view into the terminal state for rendering a single frame.

Layout

The primary layout engine for dividing terminal space using constraints and direction.

Line

A line of text, consisting of one or more Spans.

Margin

Represents spacing around rectangular areas.

Masked

A wrapper around a string that is masked when displayed.

Modifier

Modifier changes the way a piece of text is displayed.

Position

Position in the terminal coordinate system.

Rect

A rectangular area in the terminal.

Size

A simple size struct for representing dimensions in the terminal.

Span

Represents a part of a line that is contiguous and where all characters share the same style.

Style

Style lets you control the main characteristics of the displayed elements.

Terminal

An interface to interact and draw Frames on the user’s terminal.

Text

A string split over one or more lines.

Enums

Color

ANSI Color

Constraint

A constraint that defines the size of a layout element.

Direction

Defines the direction of a layout.

HorizontalAlignment

Horizontal content alignment within a layout area.

VerticalAlignment

Vertical content alignment within a layout area.

Traits

Backend

The Backend trait provides an abstraction over different terminal libraries. It defines the methods required to draw content, manipulate the cursor, and clear the terminal screen.

BlockExt

An extension trait for Block that provides some convenience methods.

FromCrossterm

A trait for converting a Crossterm type to a Ratatui type.

IntoCrossterm

A trait for converting a Ratatui type to a Crossterm type.

StatefulWidget

A StatefulWidget is a widget that can take advantage of some local state to remember things between two draw calls.

Stylize

An extension trait for styling objects.

Widget

A Widget is a type that can be drawn on a Buffer in a given Rect.

Type Aliases

Alignment

A type alias for HorizontalAlignment.