Module ratatui::prelude

A prelude for conveniently writing applications using this library.

use ratatui::prelude::*;

Aside from the main types that are used in the library, this prelude also re-exports several modules to make it easy to qualify types that would otherwise collide. E.g.:

use ratatui::{prelude::*, widgets::*};
use ratatui::widgets::{Block, Borders};

#[derive(Debug, Default, PartialEq, Eq)]
struct Line;

assert_eq!(Line::default(), Line);
assert_eq!(text::Line::default(), ratatui::text::Line::from(vec![]));

Re-exports

pub use crate::backend::CrosstermBackend;

pub use crate::backend;

pub use crate::backend::Backend;

pub use crate::buffer;

pub use crate::buffer::Buffer;

pub use crate::layout;

pub use crate::layout::Alignment;

pub use crate::layout::Constraint;

pub use crate::layout::Corner;

pub use crate::layout::Direction;

pub use crate::layout::Layout;

pub use crate::layout::Margin;

pub use crate::layout::Rect;

pub use crate::style;

pub use crate::style::Color;

pub use crate::style::Modifier;

pub use crate::style::Style;

pub use crate::style::Styled;

pub use crate::style::Stylize;

pub use crate::symbols;

pub use crate::symbols::Marker;

pub use crate::terminal;

pub use crate::terminal::Frame;

pub use crate::terminal::Terminal;

pub use crate::terminal::TerminalOptions;

pub use crate::terminal::Viewport;

pub use crate::text;

pub use crate::text::Line;

pub use crate::text::Masked;

pub use crate::text::Span;

pub use crate::text::Text;