ratatui::prelude

Trait Stylize

pub trait Stylize<'a, T>: Sized {

    // Required methods
    fn bg<C>(self, color: C) -> T
       where C: Into<Color>;
    fn fg<C>(self, color: C) -> T
       where C: Into<Color>;
    fn reset(self) -> T;
    fn add_modifier(self, modifier: Modifier) -> T;
    fn remove_modifier(self, modifier: Modifier) -> T;

    // Provided methods
    fn black(self) -> T { ... }
    fn on_black(self) -> T { ... }
    fn red(self) -> T { ... }
    fn on_red(self) -> T { ... }
    fn green(self) -> T { ... }
    fn on_green(self) -> T { ... }
    fn yellow(self) -> T { ... }
    fn on_yellow(self) -> T { ... }
    fn blue(self) -> T { ... }
    fn on_blue(self) -> T { ... }
    fn magenta(self) -> T { ... }
    fn on_magenta(self) -> T { ... }
    fn cyan(self) -> T { ... }
    fn on_cyan(self) -> T { ... }
    fn gray(self) -> T { ... }
    fn on_gray(self) -> T { ... }
    fn dark_gray(self) -> T { ... }
    fn on_dark_gray(self) -> T { ... }
    fn light_red(self) -> T { ... }
    fn on_light_red(self) -> T { ... }
    fn light_green(self) -> T { ... }
    fn on_light_green(self) -> T { ... }
    fn light_yellow(self) -> T { ... }
    fn on_light_yellow(self) -> T { ... }
    fn light_blue(self) -> T { ... }
    fn on_light_blue(self) -> T { ... }
    fn light_magenta(self) -> T { ... }
    fn on_light_magenta(self) -> T { ... }
    fn light_cyan(self) -> T { ... }
    fn on_light_cyan(self) -> T { ... }
    fn white(self) -> T { ... }
    fn on_white(self) -> T { ... }
    fn bold(self) -> T { ... }
    fn not_bold(self) -> T { ... }
    fn dim(self) -> T { ... }
    fn not_dim(self) -> T { ... }
    fn italic(self) -> T { ... }
    fn not_italic(self) -> T { ... }
    fn underlined(self) -> T { ... }
    fn not_underlined(self) -> T { ... }
    fn slow_blink(self) -> T { ... }
    fn not_slow_blink(self) -> T { ... }
    fn rapid_blink(self) -> T { ... }
    fn not_rapid_blink(self) -> T { ... }
    fn reversed(self) -> T { ... }
    fn not_reversed(self) -> T { ... }
    fn hidden(self) -> T { ... }
    fn not_hidden(self) -> T { ... }
    fn crossed_out(self) -> T { ... }
    fn not_crossed_out(self) -> T { ... }

}

An extension trait for styling objects.

For any type that implements Stylize, the provided methods in this trait can be used to style the type further. This trait is automatically implemented for any type that implements the Styled trait which e.g.: String, &str, Span, Style and many Widget types.

This results in much more ergonomic styling of text and widgets. For example, instead of writing:

let text = Span::styled("Hello", Style::default().fg(Color::Red).bg(Color::Blue));

You can write:

let text = "Hello".red().on_blue();

This trait implements a provided method for every color as both foreground and background (prefixed by on_), and all modifiers as both an additive and subtractive modifier (prefixed by not_). The reset() method is also provided to reset the style.

Examples

use ratatui_core::{
    style::{Color, Modifier, Style, Stylize},
    text::Line,
    widgets::{Block, Paragraph},
};

let span = "hello".red().on_blue().bold();
let line = Line::from(vec![
    "hello".red().on_blue().bold(),
    "world".green().on_yellow().not_bold(),
]);
let paragraph = Paragraph::new(line).italic().underlined();
let block = Block::bordered().title("Title").on_white().bold();

Required Methods

fn bg<C>(self, color: C) -> T
where C: Into<Color>,

fn fg<C>(self, color: C) -> T
where C: Into<Color>,

fn reset(self) -> T

fn add_modifier(self, modifier: Modifier) -> T

fn remove_modifier(self, modifier: Modifier) -> T

Provided Methods

fn black(self) -> T

Sets the foreground color to black.

fn on_black(self) -> T

Sets the background color to black.

fn red(self) -> T

Sets the foreground color to red.

fn on_red(self) -> T

Sets the background color to red.

fn green(self) -> T

Sets the foreground color to green.

fn on_green(self) -> T

Sets the background color to green.

fn yellow(self) -> T

Sets the foreground color to yellow.

fn on_yellow(self) -> T

Sets the background color to yellow.

fn blue(self) -> T

Sets the foreground color to blue.

fn on_blue(self) -> T

Sets the background color to blue.

fn magenta(self) -> T

Sets the foreground color to magenta.

fn on_magenta(self) -> T

Sets the background color to magenta.

fn cyan(self) -> T

Sets the foreground color to cyan.

fn on_cyan(self) -> T

Sets the background color to cyan.

fn gray(self) -> T

Sets the foreground color to gray.

fn on_gray(self) -> T

Sets the background color to gray.

fn dark_gray(self) -> T

Sets the foreground color to dark_gray.

fn on_dark_gray(self) -> T

Sets the background color to dark_gray.

fn light_red(self) -> T

Sets the foreground color to light_red.

fn on_light_red(self) -> T

Sets the background color to light_red.

fn light_green(self) -> T

Sets the foreground color to light_green.

fn on_light_green(self) -> T

Sets the background color to light_green.

fn light_yellow(self) -> T

Sets the foreground color to light_yellow.

fn on_light_yellow(self) -> T

Sets the background color to light_yellow.

fn light_blue(self) -> T

Sets the foreground color to light_blue.

fn on_light_blue(self) -> T

Sets the background color to light_blue.

fn light_magenta(self) -> T

Sets the foreground color to light_magenta.

fn on_light_magenta(self) -> T

Sets the background color to light_magenta.

fn light_cyan(self) -> T

Sets the foreground color to light_cyan.

fn on_light_cyan(self) -> T

Sets the background color to light_cyan.

fn white(self) -> T

Sets the foreground color to white.

fn on_white(self) -> T

Sets the background color to white.

fn bold(self) -> T

Adds the bold modifier.

fn not_bold(self) -> T

Removes the bold modifier.

fn dim(self) -> T

Adds the dim modifier.

fn not_dim(self) -> T

Removes the dim modifier.

fn italic(self) -> T

Adds the italic modifier.

fn not_italic(self) -> T

Removes the italic modifier.

fn underlined(self) -> T

Adds the underlined modifier.

fn not_underlined(self) -> T

Removes the underlined modifier.

fn slow_blink(self) -> T

Adds the slow_blink modifier.

fn not_slow_blink(self) -> T

Removes the slow_blink modifier.

fn rapid_blink(self) -> T

Adds the rapid_blink modifier.

fn not_rapid_blink(self) -> T

Removes the rapid_blink modifier.

fn reversed(self) -> T

Adds the reversed modifier.

fn not_reversed(self) -> T

Removes the reversed modifier.

fn hidden(self) -> T

Adds the hidden modifier.

fn not_hidden(self) -> T

Removes the hidden modifier.

fn crossed_out(self) -> T

Adds the crossed_out modifier.

fn not_crossed_out(self) -> T

Removes the crossed_out modifier.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.