ratatui::style::palette

Module material

Material design color palettes.

Represents the colors from the 2014 Material design color palettes by Google.

There are 16 palettes with accent colors, and 3 palettes without accent colors. Each palette has 10 colors, with variants from 50 to 900. The accent palettes also have 4 accent colors with variants from 100 to 700. Black and White are also included for completeness and to avoid being affected by any terminal theme that might be in use.

This module exists to provide a convenient way to use the colors from the matdesign-color crate in your application.

C50

C100

C200

C300

C400

C500

C600

C700

C800

C900

A100

A200

A400

A700

RED

PINK

PURPLE

DEEP_PURPLE

INDIGO

BLUE

LIGHT_BLUE

CYAN

TEAL

GREEN

LIGHT_GREEN

LIME

YELLOW

AMBER

ORANGE

DEEP_ORANGE

BROWN

GRAY

BLUE_GRAY

BLACK

WHITE

Example

use ratatui::style::{
    palette::material::{BLUE, RED},
    Color,
};

assert_eq!(RED.c500, Color::Rgb(244, 67, 54));
assert_eq!(BLUE.c500, Color::Rgb(33, 150, 243));

Structs

AccentedPalette

A palette of colors for use in Material design with accent colors

NonAccentedPalette

A palette of colors for use in Material design without accent colors

Constants

AMBER

BLACK

BLUE

BLUE_GRAY

BROWN

CYAN

DEEP_ORANGE

DEEP_PURPLE

GRAY

GREEN

INDIGO

LIGHT_BLUE

LIGHT_GREEN

LIME

ORANGE

PINK

PURPLE

RED

TEAL

WHITE

YELLOW