Trait ratatui::style::Styled

pub trait Styled {
    type Item;

    // Required methods
    fn style(&self) -> Style;
    fn set_style(self, style: Style) -> Self::Item;
}

A trait for objects that have a Style.

This trait enables generic code to be written that can interact with any object that has a Style. This is used by the Stylize trait to allow generic code to be written that can interact with any object that can be styled.

Required Associated Types

type Item

Required Methods

fn style(&self) -> Style

fn set_style(self, style: Style) -> Self::Item