Module ratatui::symbols::border

Structs

Set

Constants

DOUBLE

Border Set with a double line width

PLAIN

Border Set with a single line width

QUADRANT_BLOCK

QUADRANT_BOTTOM_HALF

QUADRANT_BOTTOM_LEFT

QUADRANT_BOTTOM_RIGHT

QUADRANT_INSIDE

Quadrant used for setting a border inside a block by one half cell “pixel”.

QUADRANT_LEFT_HALF

QUADRANT_OUTSIDE

Quadrant used for setting a border outside a block by one half cell “pixel”.

QUADRANT_RIGHT_HALF

QUADRANT_TOP_HALF

QUADRANT_TOP_LEFT

QUADRANT_TOP_LEFT_BOTTOM_LEFT_BOTTOM_RIGHT

QUADRANT_TOP_LEFT_BOTTOM_RIGHT

QUADRANT_TOP_LEFT_TOP_RIGHT_BOTTOM_LEFT

QUADRANT_TOP_LEFT_TOP_RIGHT_BOTTOM_RIGHT

QUADRANT_TOP_RIGHT

QUADRANT_TOP_RIGHT_BOTTOM_LEFT

QUADRANT_TOP_RIGHT_BOTTOM_LEFT_BOTTOM_RIGHT

ROUNDED

Border Set with a single line width and rounded corners

THICK

Border Set with a thick line width