ratatui::symbols

Module border

Structs

Set

Constants

DOUBLE

Border Set with a double line width

EMPTY

Empty border set

FULL

Solid border set

ONE_EIGHTH_BOTTOM_EIGHT

ONE_EIGHTH_LEFT_EIGHT

ONE_EIGHTH_RIGHT_EIGHT

ONE_EIGHTH_TALL

Tall border set based on McGugan box technique

ONE_EIGHTH_TOP_EIGHT

ONE_EIGHTH_WIDE

Wide border set based on McGugan box technique

PLAIN

Border Set with a single line width

PROPORTIONAL_TALL

Tall proportional (visually equal width and height) border with using set of quadrants.

PROPORTIONAL_WIDE

Wide proportional (visually equal width and height) border with using set of quadrants.

QUADRANT_BLOCK

QUADRANT_BOTTOM_HALF

QUADRANT_BOTTOM_LEFT

QUADRANT_BOTTOM_RIGHT

QUADRANT_INSIDE

Quadrant used for setting a border inside a block by one half cell “pixel”.

QUADRANT_LEFT_HALF

QUADRANT_OUTSIDE

Quadrant used for setting a border outside a block by one half cell “pixel”.

QUADRANT_RIGHT_HALF

QUADRANT_TOP_HALF

QUADRANT_TOP_LEFT

QUADRANT_TOP_LEFT_BOTTOM_LEFT_BOTTOM_RIGHT

QUADRANT_TOP_LEFT_BOTTOM_RIGHT

QUADRANT_TOP_LEFT_TOP_RIGHT_BOTTOM_LEFT

QUADRANT_TOP_LEFT_TOP_RIGHT_BOTTOM_RIGHT

QUADRANT_TOP_RIGHT

QUADRANT_TOP_RIGHT_BOTTOM_LEFT

QUADRANT_TOP_RIGHT_BOTTOM_LEFT_BOTTOM_RIGHT

ROUNDED

Border Set with a single line width and rounded corners

THICK

Border Set with a thick line width