Module ratatui::widgets

widgets is a collection of types that implement Widget or StatefulWidget or both.

All widgets are implemented using the builder pattern and are consumable objects. They are not meant to be stored but used as commands to draw common figures in the UI.

The available widgets are:

Block: a basic widget that draws a block with optional borders, titles and styles.

BarChart: displays multiple datasets as bars with optional grouping.

calendar::Monthly: displays a single month.

Canvas: draws arbitrary shapes using drawing characters.

Chart: displays multiple datasets as a lines or scatter graph.

Clear: clears the area it occupies. Useful to render over previously drawn widgets.

Gauge: displays progress percentage using block characters.

LineGauge: display progress as a line.

List: displays a list of items and allows selection.

Paragraph: displays a paragraph of optionally styled and wrapped text.

Scrollbar: displays a scrollbar.

Sparkline: display a single data set as a sparkline.

Table: displays multiple rows and columns in a grid and allows selection.

Tabs: displays a tab bar and allows selection.

Re-exports

pub use self::block::Block;

pub use self::block::BorderType;

pub use self::block::Padding;

Modules

block

Elements related to the Block base widget.

calendar

A simple calendar widget. (feature: widget-calendar)

canvas

Structs

Axis

An X or Y axis for the chart widget

Bar

A bar to be shown by the BarChart widget.

BarChart

A chart showing values as bars.

BarGroup

A group of bars to be shown by the Barchart.

Borders

Bitflags that can be composed to set the visible borders essentially on the block widget.

Cell

A Cell contains the Text to be displayed in a Row of a Table.

Chart

A widget to plot one or more dataset in a cartesian coordinate system

Clear

A widget to clear/reset a certain area to allow overdrawing (e.g. for popups).

Dataset

A group of data points

Gauge

A widget to display a progress bar.

LineGauge

A compact widget to display a progress bar over a single thin line.

List

A widget to display several items among which one can be selected (optional)

ListItem

ListState

Paragraph

A widget to display some text.

Row

Holds data to be displayed in a Table widget.

Scrollbar

A widget to display a scrollbar

ScrollbarState

A struct representing the state of a Scrollbar widget.

Sparkline

Widget to render a sparkline over one or more lines.

Table

A widget to display data in formatted columns.

TableState

Tabs

A widget that displays a horizontal set of Tabs with a single tab selected.

Wrap

Describes how to wrap text across lines.

Enums

GraphType

Used to determine which style of graphing to use

HighlightSpacing

This option allows the user to configure the “highlight symbol” column width spacing

RenderDirection

ScrollDirection

An enum representing the direction of scrolling in a Scrollbar widget.

ScrollbarOrientation

Scrollbar Orientation

Traits

StatefulWidget

A StatefulWidget is a widget that can take advantage of some local state to remember things between two draw calls.

Widget

Base requirements for a Widget