Module ratatui::widgets

widgets is a collection of types that implement Widget or StatefulWidget or both.

All widgets are implemented using the builder pattern and are consumable objects. They are not meant to be stored but used as commands to draw common figures in the UI.

The available widgets are:

Block

Tabs

List

Table

Paragraph

Chart

BarChart

Gauge

Sparkline

[calendar::Monthly]

Clear

Re-exports

pub use self::block::Block;

pub use self::block::BorderType;

pub use self::block::Padding;

pub use self::scrollbar::ScrollDirection;

pub use self::scrollbar::Scrollbar;

pub use self::scrollbar::ScrollbarOrientation;

pub use self::scrollbar::ScrollbarState;

Modules

block

canvas

scrollbar

Structs

Axis

An X or Y axis for the chart widget

Bar

represent a bar to be shown by the Barchart

BarChart

Display multiple bars in a single widgets

BarGroup

represent a group of bars to be shown by the Barchart

Borders

Bitflags that can be composed to set the visible borders essentially on the block widget.

Cell

A Cell contains the Text to be displayed in a Row of a Table.

Chart

A widget to plot one or more dataset in a cartesian coordinate system

Clear

A widget to clear/reset a certain area to allow overdrawing (e.g. for popups).

Dataset

A group of data points

Gauge

A widget to display a task progress.

LineGauge

A compact widget to display a task progress over a single line.

List

A widget to display several items among which one can be selected (optional)

ListItem

ListState

Paragraph

A widget to display some text.

Row

Holds data to be displayed in a Table widget.

Sparkline

Widget to render a sparkline over one or more lines.

Table

A widget to display data in formatted columns.

TableState

Tabs

A widget to display available tabs in a multiple panels context.

Wrap

Describes how to wrap text across lines.

Enums

GraphType

Used to determine which style of graphing to use

RenderDirection

Traits

StatefulWidget

A StatefulWidget is a widget that can take advantage of some local state to remember things between two draw calls.

Widget

Base requirements for a Widget