Struct ratatui::widgets::BarChart

pub struct BarChart<'a> { /* private fields */ }

Display multiple bars in a single widgets

Examples

The following example creates a BarChart with two groups of bars. The first group is added by an array slice (&[(&str, u64)]). The second group is added by a slice of Groups (&BarGroup).

BarChart::default()
    .block(Block::default().title("BarChart").borders(Borders::ALL))
    .bar_width(3)
    .bar_gap(1)
    .group_gap(3)
    .bar_style(Style::new().yellow().on_red())
    .value_style(Style::new().red().bold())
    .label_style(Style::new().white())
    .data(&[("B0", 0), ("B1", 2), ("B2", 4), ("B3", 3)])
    .data(BarGroup::default().bars(&[Bar::default().value(10), Bar::default().value(20)]))
    .max(4);

Implementations

impl<'a> BarChart<'a>

pub fn data(self, data: impl Into<BarGroup<'a>>) -> BarChart<'a>

Add group of bars to the BarChart

Examples

The following example creates a BarChart with two groups of bars. The first group is added by an array slice (&[(&str, u64)]). The second group is added by a BarGroup instance.

BarChart::default()
       .data(&[("B0", 0), ("B1", 2), ("B2", 4), ("B3", 3)])
       .data(BarGroup::default().bars(&[Bar::default().value(10), Bar::default().value(20)]));

pub fn block(self, block: Block<'a>) -> BarChart<'a>

pub fn max(self, max: u64) -> BarChart<'a>

pub fn bar_style(self, style: Style) -> BarChart<'a>

pub fn bar_width(self, width: u16) -> BarChart<'a>

pub fn bar_gap(self, gap: u16) -> BarChart<'a>

pub fn bar_set(self, bar_set: Set) -> BarChart<'a>

pub fn value_style(self, style: Style) -> BarChart<'a>

pub fn label_style(self, style: Style) -> BarChart<'a>

pub fn group_gap(self, gap: u16) -> BarChart<'a>

pub fn style(self, style: Style) -> BarChart<'a>