Trait ratatui::widgets::Widget

pub trait Widget {
    // Required method
    fn render(self, area: Rect, buf: &mut Buffer);
}

Base requirements for a Widget

Required Methods

fn render(self, area: Rect, buf: &mut Buffer)

Draws the current state of the widget in the given buffer. That is the only method required to implement a custom widget.